Title: 낮은 지점들, Time: 0ms, Memory: 0MB, Status: Passed - Codetree

+++ 260430/낮은 지점들/lowest-points.java
@@ -6,11 +6,11 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(br.readLine());
-        HashMap<Integer, Integer> hm = new HashMap<>();
+        HashMap<Long, Long> hm = new HashMap<>();
         for(int i=0; i<N; i++){
             StringTokenizer st = new StringTokenizer(br.readLine());
-            int x = Integer.parseInt(st.nextToken());
-            int y = Integer.parseInt(st.nextToken());
+            long x = Long.parseLong(st.nextToken());
+            long y = Long.parseLong(st.nextToken());
 
             if(hm.containsKey(x)) {
                 if(hm.get(x) > y) {
@@ -21,8 +21,8 @@ public class Main {
             }
         }
 
-        int sum = 0;
-        for(int num : hm.values()) {
+        long sum = 0;
+        for(long num : hm.values()) {
             sum+=num;
         }
